+++ src/common/assets/images/imageFile.ts
@@ -1,26 +1,42 @@
 import sampleImage from "./sample.jpeg";
-export { sampleImage };
-export const stellaMcCartney =
+
+const stellaMcCartney =
   "https://hrd-live.cdn.scayle.cloud/images/ac65e8826d836a704ea585786728df44.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const bottegaVeneta =
+const bottegaVeneta =
   "https://hrd-live.cdn.scayle.cloud/images/9b59f832e8f9be188f0396c17084432f.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const sir =
+const sir =
   "https://hrd-live.cdn.scayle.cloud/images/c17e195f251364ae2ea905b7deb5bddc.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const valentinoGaravani =
+const valentinoGaravani =
   "https://hrd-live.cdn.scayle.cloud/images/4ca373ac3fb36a2f703997e9e5b7bec1.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const leoLin =
+const leoLin =
   "https://hrd-live.cdn.scayle.cloud/images/e42f69d497b907c81824cf937c6b885e.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const lieStudio =
+const lieStudio =
   "https://hrd-live.cdn.scayle.cloud/images/13548022f6f52657836b1cf3fefaed0b.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const retori =
+const retori =
   "https://hrd-live.cdn.scayle.cloud/images/d0e6da25c083b6195c5ea4b91b73f724.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const leMondeBeryl =
+const leMondeBeryl =
   "https://hrd-live.cdn.scayle.cloud/images/0a13c6346513efedbd0d5c85f8105654.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const jackuemus =
+const jackuemus =
   "https://hrd-live.cdn.scayle.cloud/images/6ae053496878e0f3e740c5729e49e919.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const christianLouboutin =
+const christianLouboutin =
   "https://hrd-live.cdn.scayle.cloud/images/9d5d19cbf82e0df14bf89f7b667c2789.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const vivienneWestwood =
+const vivienneWestwood =
   "https://hrd-live.cdn.scayle.cloud/images/87e4fd023b237f3cd96676255d0669f3.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
-export const chole =
+const chole =
   "https://hrd-live.cdn.scayle.cloud/images/40b25465416dd811479dd09f58693e5b.jpg?brightness=1&width=960&height=1280&quality=75&bg=ffffff";
+
+export {
+  bottegaVeneta,
+  chole,
+  christianLouboutin,
+  jackuemus,
+  leMondeBeryl,
+  leoLin,
+  lieStudio,
+  retori,
+  sampleImage,
+  sir,
+  stellaMcCartney,
+  valentinoGaravani,
+  vivienneWestwood,
+};
